Moved TerminationScope and TerminationSignal to core
This commit moves TerminationScope and TerminationSignal to core, to
resolve a cyclic dependency.

See: gh-1024
